+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-# 0.0.1 - A sweet treat
+# 0.0.2 - A dipping good time
 
-* Initial release
-* Party Rings!
+* Fix dipper not dipping
+* Make dipper rotatable with wrench
+* Give dipper texture a slight texture change to see which way its facing
+* Make dipper mine-able with a pickaxe